src/requirements.txt
pyyaml==6.0.2
tqdm==4.66.4
numpy==2.1.1
sentence-transformers==3.0.1
requests==2.32.3
